train model with Pre-trained BERT model
1. Tokenization
Our first step is to tokenize the sentences -- break them up into word and subwords in the format BERT is comfortable with.
2. Padding
After tokenization, tokenized is a list of sentences -- each sentences is represented as a list of tokens. We want BERT to process our examples all at once (as one batch). It's just faster that way. For that reason, we need to pad all lists to the same size, so we can represent the input as one 2-d array, rather than a list of lists (of different lengths).
3. Masking
If we directly send padded to BERT, that would slightly confuse it. We need to create another variable to tell it to ignore (mask) the padding we've added when it's processing its input.
4. The model() function runs our sentences through BERT. The results of the processing will be returned into last_hidden_states.
We'll save those in the features variable, as they'll serve as the features to our logitics regression model.
The labels indicating which sentence is positive and negative now go into the labels variable
5. Split our datset into a training set and testing set 
6. Train the LogisticRegression model
7. Show score how well our model does in classifying sentences